Fixed a bug in st7789_set_pixel
Pixel management was not correct, as the original code used a byte-level access
to modify pixel RGB components, whereas ESP32 expects dword-level access.

The code has been modified to take care of that, and now relies on uint32_t values
to set pixel RGB components. This is more complex, but performance is quite the
same as we rely on a switch to handle the 4 possible cases (at least, I hope so).